requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.1
djangorestframework==3.11.1
nodejs==0.1.1
npm==0.1.1
optional-django==0.1.0
pytz==2020.1
sqlparse==0.3.1
